+++ fs/jffs2/jffs2_1pass.c
@@ -174,7 +174,7 @@ static int read_nand_cached(u32 off, u32 size, u_char *buf)
 {
 	struct mtdids *id = current_part->dev->id;
 	u32 bytes_read = 0;
-	size_t retlen;
+	uint64_t retlen;
 	int cpy_bytes;
 
 	while (bytes_read < size) {
